Weird Main stuffs
'Nuff said.
